+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/service/admin/staff/SaveJudgeException.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.ethos.replacement.docmosis.service.admin.staff.judge;
+package uk.gov.hmcts.ethos.replacement.docmosis.service.admin.staff;
 
 public class SaveJudgeException extends RuntimeException {
     public SaveJudgeException(String message) {
