+++ src/test/java/uk/gov/hmcts/ethos/replacement/docmosis/service/admin/staff/JudgeServiceTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.ethos.replacement.docmosis.service.admin.staff.judge;
+package uk.gov.hmcts.ethos.replacement.docmosis.service.admin.staff;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -10,12 +10,11 @@ import uk.gov.hmcts.ethos.replacement.docmosis.domain.repository.JudgeRepository
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.never;
-
 
 class JudgeServiceTest {
     private JudgeRepository judgeRepository;
@@ -36,7 +35,8 @@ class JudgeServiceTest {
 
     @Test
     void shouldSaveJudge() {
-        when(judgeRepository.existsByCodeAndTribunalOffice(judgeCode, TribunalOffice.valueOfOfficeName(tribunalOffice))).thenReturn(false);
+        when(judgeRepository.existsByCodeAndTribunalOffice(judgeCode, TribunalOffice.valueOfOfficeName(tribunalOffice)))
+                .thenReturn(false);
         when(judgeRepository.existsByNameAndTribunalOffice(judgeName,
                 TribunalOffice.valueOfOfficeName(tribunalOffice))).thenReturn(false);
 
@@ -46,21 +46,24 @@ class JudgeServiceTest {
 
     @Test
     void shouldReturnErrorIfJudgeWithSameCodeAndOfficeExists() {
-        when(judgeRepository.existsByCodeAndTribunalOffice(judgeCode,TribunalOffice.valueOfOfficeName(tribunalOffice))).thenReturn(true);
+        when(judgeRepository.existsByCodeAndTribunalOffice(judgeCode, TribunalOffice.valueOfOfficeName(tribunalOffice)))
+                .thenReturn(true);
         assertThrows(SaveJudgeException.class, () -> addJudgeService.saveJudge(adminData));
         verify(judgeRepository, never()).save(judge);
     }
 
     @Test
     void shouldReturnErrorIfJudgeWithSameNameAndOfficeExists() {
-        when(judgeRepository.existsByCodeAndTribunalOffice(judgeCode,TribunalOffice.valueOfOfficeName(tribunalOffice))).thenReturn(false);
+        when(judgeRepository.existsByCodeAndTribunalOffice(judgeCode, TribunalOffice.valueOfOfficeName(tribunalOffice)))
+                .thenReturn(false);
         when(judgeRepository.existsByNameAndTribunalOffice(judgeName,
                 TribunalOffice.valueOfOfficeName(tribunalOffice))).thenReturn(true);
         assertThrows(SaveJudgeException.class, () -> addJudgeService.saveJudge(adminData));
         verify(judgeRepository, never()).save(judge);
     }
 
-    private AdminData createAdminData(String judgeCode, String judgeName, String tribunalOffice, String employmentStatus) {
+    private AdminData createAdminData(String judgeCode, String judgeName, String tribunalOffice,
+                                      String employmentStatus) {
         AdminData adminData = new AdminData();
         adminData.setJudgeCode(judgeCode);
         adminData.setJudgeName(judgeName);
